requirements.txt
selenium
pytest
python-dotenv
webdriver-manager